+++ besu/src/main/java/org/hyperledger/besu/cli/BesuCommand.java
@@ -1287,8 +1287,8 @@ public class BesuCommand implements DefaultCommandValues, Runnable {
   @CommandLine.Option(
       names = {"--rpc-max-logs-range"},
       description =
-          "Specifies the maximum number of blocks to retrieve logs from via RPC. Value must be > 0")
-  private final Long rpcMaxLogsRange = 0L;
+          "Specifies the maximum number of blocks to retrieve logs from via RPC. Must be >=0. 0 specifies no limit  (default: ${DEFAULT-VALUE})")
+  private final Long rpcMaxLogsRange = 1000L;
 
   @Mixin private P2PTLSConfigOptions p2pTLSConfigOptions;